Implement the vehicle list window (#878)
* Implementing open function

* Implement widgets, stub events.

* Implement company switching (mouse down, dropdown events)

* Implement update and tooltip events.

* Implement remaining events

* Implement tab and sort mode switching.

* Implement prepareDraw and draw events.

* Implement VehicleList::drawTabs

* Implement refreshVehicleList

* WIP updateVehicleList

* Fix vehicle type checking

* Implement orderByProfit

* Implement orderByAge

* Implement orderByReliability

* Implement onScrollMouseDown

* Implement drawScroll event

* WIP onScrollMouseOver

* Move generateCargoTotalString into VehicleHead

* Finish onScrollMouseOver

* Use VehicleType type in more places.

+++ src/OpenLoco/Input/MouseInput.cpp
@@ -772,7 +772,7 @@ namespace OpenLoco::Input
                                 if (veh != nullptr)
                                 {
                                     auto head = EntityManager::get<Vehicles::VehicleHead>(veh->getHead());
-                                    Ui::Windows::VehicleList::open(head->owner, static_cast<uint8_t>(head->vehicleType));
+                                    Ui::Windows::VehicleList::open(head->owner, head->vehicleType);
                                 }
                                 break;
                             }
